Added 'CodeStores' class
'CodeStores' class aim to stores all the generated code and provides
common functionalities for code generator.
